Moved status bar to top, changed offlineimap UI.
Moved mutt status bar to the top of the display as it's easier to see what
mailbaox is currently selected that way. This puts the list of command options
at the bottow of the display, which work too.

Tried `quiet` as the offlineimap UI -- too litte information. Reverted to
blinkenlights for now.