Fixed command prompt to work on OS X and Linux
Made changes to prompt_command function to only echo the information I want,
and PS1 now is only the "$ " prompt on a newline. Had to change away from the
PS1 flags for user, host, and working directory, and use id, hostname, and pwd
for that information.

All of this to get git branch information to properly update on Linux.