Revised how snippets are filled in
The replacement targets in the snippets all start and end with two
underscores ( __ ), this allows the use of a Vim find command with
regular expressions to highlight them all after the snippet is included
in the target file. Using `cw` to change the current word, and `n` to go
to the next target makes for a nice way to fill in the snippet.

The two Chef snippets were updated to use the new double underscore
naming of targets