Added override-redirect to polybar config
This setting seems to be the way to allow windows to redraw correctly
when polybar is drawn on the screen. Since I have two polybars, I needed
this. The time of when the screen contents were shown and when the
polybr was shown sometimes caused overlap between the two elements.
Solved thanks to this posting on unixporn:
http://reddit.com/r/unixporn/comments/6t92gr/how_to_get_i3_to_respect_polybar_height